[ch155] Restructure files
