+++ src/components/Navbar/index.tsx
@@ -1,12 +1,12 @@
 import React from 'react'
-import NavbarLogo from './Logo'
+import Logo from '../Logo'
 import NavbarMenu from './Menu'
 import NavbarTool from './Tool'
 
 const Navbar = () => {
   return (
     <nav className="navbar">
-      <NavbarLogo />
+      <Logo />
       <NavbarMenu />
       <NavbarTool />
     </nav>
