+++ src/enum/ComponentType.ts
@@ -1,0 +1,4 @@
+export enum ComponentType {
+  Header,
+  Footer
+}
