+++ src/styles/components/__components-dir.scss
@@ -23,3 +23,4 @@
 @import './home-category';
 @import './dropdown';
 @import './home-search';
+@import './logo';
