+++ src/styles/components/_logo.scss
@@ -1,0 +1,16 @@
+.logo {
+  color: $dark;
+  font-size: rem(24px);
+  font-weight: bold;
+  letter-spacing: rem(4px);
+  text-align: left;
+  text-decoration: none;
+
+  @include respond-above('md') {
+    font-size: rem(32px);
+  }
+
+  @include respond-above('lg') {
+    font-size: rem(36px);
+  }
+}
